A script for windows, to automatically rename all the hardlinks associated with a the file located in a master folder. The script does this in a following manner,
Run a for loop, which access each file of the directory in which the script is located.
Extracts all the hardlinks located in the computer associated with each file.
Creates a renaming command and saves it a temporary file.

In the end of the loop runs the renaming script then deletes it.